\name{get.recruitment}
\alias{get.recruitment}
\title{Calculate fish recruitment}
\usage{
  get.recruitment(Fvars, SSB)
}
\arguments{
  \item{Fvars}{Named list of fish simulation variables
  created by call to \code{\link{set.fish.pars}}}

  \item{SSB}{Spawning biomass}
}
\description{
  Calculate fish recruitment
}
\author{
  Paul Conn
}

