\docType{data}
\name{HFYS_appendix_C}
\alias{HFYS_appendix_C}
\title{Data from Holmes et al. (2007) Appendix C}
\description{
  Data from Holmes et al. (2007) Appendix C
}
\references{
  Holmes , E., Fritz, L., York, A., and Sweeny, K. (2007)
  Age-structured modeling reveals long-term declines in the
  natality of western Steller sea lions. Ecological
  Applications 17:2214-2232. Repreint available at:
  \url{http://alaskafisheries.noaa.gov/protectedresources/stellers/esa/biop/draft/holmes_etal_2007_ssl_agestructure.pdf}
}
\keyword{datasets}

