\name{init.fish}
\alias{init.fish}
\title{Initialize fish abundance}
\usage{
  init.fish(Fvars, isite)
}
\arguments{
  \item{Fvars}{Named list of fish simulation variables
  created by call to \code{\link{set.fish.pars}}}

  \item{isite}{Integer site number}
}
\description{
  Initialize fish abundance
}
\author{
  Paul Conn
}

