\name{obj_fun_SSLfish}
\alias{obj_fun_SSLfish}
\title{Objective function for simulation calibration}
\usage{
  obj_fun_SSLfish(pars, fixed.vars)
}
\arguments{
  \item{pars}{Simulation parameters to calibrate}

  \item{fixed.vars}{List of simulation parameters that are
  fixed}
}
\description{
  Objective function for simulation calibration
}
\details{
  More details to come? ....
}
\author{
  Paul Conn
}

