\name{project.fish.1yr}
\alias{project.fish.1yr}
\title{Project fish population forward 1 year}
\usage{
  project.fish.1yr(Fvars, Svars, FLvars, n.sites, iyr)
}
\arguments{
  \item{Fvars}{Fish population variables list created by
  call to \code{\link{set.fish.pars}}}

  \item{Svars}{Sea lion population variables list created
  by call to \code{\link{set.ssl.pars}}}

  \item{FLvars}{Fishing fleets variables list created by
  call to \code{\link{set.fleet.pars}}}

  \item{n.sites}{Number of sites in the simulation study}

  \item{iyr}{Current year in the simulation}
}
\description{
  Project fish population forward 1 year
}
\author{
  Paul Conn
}

