\name{set.fish.pars}
\alias{set.fish.pars}
\title{Initialize fish simulation variables}
\usage{
  set.fish.pars(n.sites, n.yrs)
}
\arguments{
  \item{n.sites}{Number of sites in the simulation}

  \item{n.yrs}{Number of total years in the simulation}
}
\description{
  Initialize fish simulation variables
}
\details{
  Add more details here...
}
\author{
  Paul Conn
}

