\name{set.fleet.pars}
\alias{set.fleet.pars}
\title{Initialize fishing fleet simulation variables}
\usage{
  set.fleet.pars(n.sites, n.yrs, burnin)
}
\arguments{
  \item{n.sites}{Number of sites in the simulation}

  \item{n.yrs}{Number of total years in the simulation}

  \item{burnin}{Number of years in the simulation before
  'fishing' begins.}
}
\description{
  Initialize fishing fleet simulation variables
}
\details{
  Add more details here...
}
\author{
  Paul Conn
}

