\docType{package}
\name{SSLfish-package}
\alias{SSLfish}
\alias{SSLfish-package}
\title{R package for simulating effects of fisheries removal of prey on Steller sea lion populations}
\description{
  This package simultaneously simulates a sea lion
  population, prey base, and fisheries removal of prey. The
  primary goal is to investiagte the validity of approaches
  used previously in the literature to assess the effects
  of fishing on the abundance of Steller sea lions.
}
\details{
  \tabular{ll}{ Package: \tab SSLfish \cr Type: \tab
  Package\cr Version: \tab 0.1\cr Date: \tab 2013-03-06\cr
  License: \tab Unlimited\cr LazyLoad: \tab yes\cr }
}
\author{
  Devin Johnson and Paul Conn

  Maintainer: Devin S. Johnson <devin.johnson@noaa.gov>
}

