\name{unfished.bpr}
\alias{unfished.bpr}
\title{Compute unfished spawning biomass per recruit}
\usage{
  unfished.bpr(Fvars)
}
\arguments{
  \item{Fvars}{Named list of fish simulation variables
  created by call to \code{\link{set.fish.pars}}}
}
\description{
  Compute unfished spawning biomass per recruit
}
\author{
  Paul Conn
}

